requirements.txt
torch
torchvision
torchaudio
diffusers[torch]==0.30.0
transformers==4.44.0
accelerate==0.44.0
huggingface_hub==0.25.0
opencv-python
numpy
Pillow
fastapi
uvicorn
python-multipart
bitsandbytes
xformers
scikit-image
trimesh
open3d
controlnet-aux
pytesseract
